added dbcommons to existing funcs()

+++ models/scoreboard-model.js
@@ -1,4 +1,5 @@
 const mongoose = require("mongoose");
+const dbcommons = require("./dbcommons");
 
 const scoreboardSchema = new mongoose.Schema({
     playerId: {type: String, required: true, unique: true},
@@ -11,10 +12,12 @@ const scoreboardSchema = new mongoose.Schema({
 const Scoreboard = mongoose.model('Scoreboard', scoreboardSchema);
 
 exports.getTopTen = function() {
+    if (!dbcommons.isDBConnected()) throw dbcommons.databaseError;
     return Scoreboard.find().sort({ highscore: -1 }).limit(10);
 };
 
 exports.upsertScore = function(playerId, username, newScore) {
+    if (!dbcommons.isDBConnected()) throw dbcommons.databaseError;
     return Scoreboard.findOneAndUpdate(
         { playerId: playerId },
         {
@@ -27,9 +30,11 @@ exports.upsertScore = function(playerId, username, newScore) {
 };
 
 exports.deleteScore = function(playerId) {
+    if (!dbcommons.isDBConnected()) throw dbcommons.databaseError;
     return Scoreboard.deleteOne({ playerId: playerId });
 };
 
 exports.getFriendScores = function(playerIds) {
+    if (!dbcommons.isDBConnected()) throw dbcommons.databaseError;
     return Scoreboard.find({playerId: {$in: playerIds}}).sort({highscore: -1});
 };
